backend/requirements/prod.txt
-r common.txt
django-storages[azure]
gunicorn
psycopg2-binary